fix(docs): keep each runtime provider in its own client module
collapsing the four into one "use client" module coupled their import graphs: every export of a client module is reachable, so importing the playground provider pulled the whole file in, and with it react-devtools and the docs toolkit that transitively drags react-generative-ui, the weather widget and the notepad tool ui. the playground's client graph carried none of that before.

the deduplication lives in chat-runtime.ts, not in the co-location, so the providers move back to a module each and the line savings stay.

sendAutomaticallyWhen also has to stay per surface rather than becoming unconditional: the playground never had it, and it registers a client-executed update_config tool, so firing it would turn every config update from the builder into an extra round trip.
